+++ src/quri_parts_oqtopus/backend/__init__.py
@@ -1,14 +1,22 @@
 from .sampling import (  # noqa: I001
     OqtopusSamplingBackend,
+)
+
+from .data.sampling import (
     OqtopusSamplingJob,
     OqtopusSamplingResult,
 )
+
 from .device import OqtopusDevice, OqtopusDeviceBackend
 from .estimation import (
     OqtopusEstimationBackend,
+)
+
+from .data.estimation import (
     OqtopusEstimationJob,
     OqtopusEstimationResult,
 )
+
 from .sse import OqtopusSseBackend
 
 from .config import OqtopusConfig
