+++ src/quri_parts_oqtopus/backend/estimation.py
@@ -1,9 +1,3 @@
-import json
-import os
-import time
-from datetime import datetime
-from typing import Any
-
 from quri_parts.backend import BackendError
 from quri_parts.circuit import NonParametricQuantumCircuit
 from quri_parts.core.operator import Operator
@@ -12,12 +6,9 @@ from quri_parts.openqasm.circuit import convert_to_qasm_str
 from quri_parts_oqtopus.backend.config import (
     OqtopusConfig,
 )
-from quri_parts_oqtopus.backend.utils import DateTimeEncoder
+from quri_parts_oqtopus.backend.data.estimation import OqtopusEstimationJob
+from quri_parts_oqtopus.backend.job_backend_base import OqtopusJobBackendBase
 from quri_parts_oqtopus.rest import (
-    ApiClient,
-    Configuration,
-    JobApi,
-    JobsJobDef,
     JobsOperatorItem,
     JobsSubmitJobInfo,
     JobsSubmitJobRequest,
@@ -26,369 +17,7 @@ from quri_parts_oqtopus.rest import (
 JOB_FINAL_STATUS = ["succeeded", "failed", "cancelled"]
 
 
-class OqtopusEstimationResult:
-    """A result of a estimation job.
-
-    Args:
-        result: A result of dict type.
-
-    Raises:
-        ValueError: If ``estimation`` does not exist in result.
-
-    Examples:
-        An example of a dict of result is as below:
-
-        .. code-block::
-
-            {
-                "estimation": {
-                    "exp_value": 2.0,
-                    "stds": 1.1
-                }
-            }
-
-        ``exp_value`` represents the expectation value.
-        In the above case, ``exp_value`` is defined as a real number ``2.0``.
-        ``stds`` represents the standard deviation value.
-
-    """
-
-    def __init__(self, result: dict[str, Any]) -> None:
-        super().__init__()
-
-        self._result = result
-
-    @property
-    def exp_value(self) -> float | None:
-        """Returns the expectation value."""
-        return self._result.get("exp_value")
-
-    @property
-    def stds(self) -> float | None:
-        """Returns the  standard deviation."""
-        return self._result.get("stds")
-
-    def __repr__(self) -> str:
-        """Return a string representation.
-
-        Returns:
-            str: A string representation.
-
-        """
-        return str(self._result)
-
-
-class OqtopusEstimationJob:  # noqa: PLR0904
-    """A job for a estimation.
-
-    Args:
-        job: A result of dict type.
-        job_api: A result of dict type.
-
-    Raises:
-        ValueError: If ``job`` or ``job_api`` is None.
-
-    """
-
-    def __init__(self, job: JobsJobDef, job_api: JobApi) -> None:
-        super().__init__()
-
-        if job is None:
-            msg = "'job' should not be None"
-            raise ValueError(msg)
-        self._job: JobsJobDef = job
-
-        if job_api is None:
-            msg = "'job_api' should not be None"
-            raise ValueError(msg)
-        self._job_api: JobApi = job_api
-
-    @property
-    def job_id(self) -> str:
-        """The id of the job.
-
-        Returns:
-            str: The id of the job.
-
-        """
-        return self._job.job_id
-
-    @property
-    def name(self) -> str:
-        """The name of the job.
-
-        Returns:
-            str: The name of the job.
-
-        """
-        return self._job.name
-
-    @property
-    def description(self) -> str:
-        """The description of the job.
-
-        Returns:
-            str: The description of the job.
-
-        """
-        return self._job.description
-
-    @property
-    def job_type(self) -> str:
-        """The job type of the job.
-
-        Returns:
-            str: The job type of the job.
-
-        """
-        return self._job.job_type
-
-    @property
-    def status(self) -> str:
-        """The status of the job.
-
-        Returns:
-            str: The status of the job.
-
-        """
-        return self._job.status
-
-    @property
-    def device_id(self) -> str:
-        """The device id of the job.
-
-        Returns:
-            str: The device id of the job.
-
-        """
-        return self._job.device_id
-
-    @property
-    def shots(self) -> int:
-        """The shots of the job.
-
-        Returns:
-            int: The shots of the job.
-
-        """
-        return self._job.shots
-
-    @property
-    def job_info(self) -> dict:
-        """The detail information of the job.
-
-        Returns:
-            dict: The detail information of the job.
-
-        """
-        return self._job.job_info.to_dict()
-
-    @property
-    def transpiler_info(self) -> dict:
-        """The transpiler info of the job.
-
-        Returns:
-            dict: The transpiler info of the job.
-
-        """
-        if self._job.transpiler_info:
-            return json.loads(self._job.transpiler_info)
-        return {}
-
-    @property
-    def simulator_info(self) -> dict:
-        """The simulator info of the job.
-
-        Returns:
-            dict: The simulator info of the job.
-
-        """
-        if self._job.simulator_info:
-            return json.loads(self._job.simulator_info)
-        return {}
-
-    @property
-    def mitigation_info(self) -> dict:
-        """The mitigation info of the job.
-
-        Returns:
-            dict: The mitigation info of the job.
-
-        """
-        if self._job.mitigation_info:
-            return json.loads(self._job.mitigation_info)
-        return {}
-
-    @property
-    def execution_time(self) -> float:
-        """The execution time of the job.
-
-        Returns:
-            float: The execution time of the job.
-
-        """
-        return self._job.execution_time
-
-    @property
-    def submitted_at(self) -> datetime:
-        """The `submitted_at` of the job.
-
-        Returns:
-            datetime: The `submitted_at` of the job.
-
-        """
-        return self._job.submitted_at
-
-    @property
-    def ready_at(self) -> datetime:
-        """The `ready_at` of the job.
-
-        Returns:
-            datetime: The `ready_at` of the job.
-
-        """
-        return self._job.ready_at
-
-    @property
-    def running_at(self) -> datetime:
-        """The `running_at` of the job.
-
-        Returns:
-            datetime: The `running_at` of the job.
-
-        """
-        return self._job.running_at
-
-    @property
-    def ended_at(self) -> datetime:
-        """The `ended_at` of the job.
-
-        Returns:
-            datetime: The `ended_at` of the job.
-
-        """
-        return self._job.ended_at
-
-    def refresh(self) -> None:
-        """Retrieve the latest job information from OQTOPUS Cloud.
-
-        Raises:
-            BackendError: If job cannot be found or if an authentication error occurred
-                or timeout occurs, etc.
-
-        """
-        try:
-            self._job = self._job_api.get_job(self._job.job_id)
-        except Exception as e:
-            msg = "To refresh job is failed."
-            raise BackendError(msg) from e
-
-    def wait_for_completion(
-        self, timeout: float | None = None, wait: float = 10.0
-    ) -> JobsJobDef | None:
-        """Wait until the job progress to the end.
-
-        Calling this function waits until the job progress to the end such as
-        ``succeeded`` or ``failed``, ``cancelled``.
-
-        Args:
-            timeout: The number of seconds to wait for job.
-            wait: Time in seconds between queries.
-
-        Returns:
-            JobsJobDef | None: If a timeout occurs, it returns None. Otherwise, it
-                returns the Job.
-
-        """
-        start_time = time.time()
-        self.refresh()
-        while self._job.status not in JOB_FINAL_STATUS:
-            # check timeout
-            elapsed_time = time.time() - start_time
-            if timeout is not None and elapsed_time >= timeout:
-                return None
-
-            # sleep and get job
-            time.sleep(wait)
-            self.refresh()
-
-        return self._job
-
-    def result(
-        self, timeout: float | None = None, wait: float = 10.0
-    ) -> OqtopusEstimationResult:
-        """Wait until the job progress to the end and returns the result of the job.
-
-        If the status of job is not ``succeeded`` or ``failed``, or ``cancelled``,
-        the job is retrieved from OQTOPUS Cloud at intervals of ``wait`` seconds.
-        If the job does not progress to the end after ``timeout`` seconds,
-        raise :class:`BackendError`.
-
-        Args:
-            timeout: The number of seconds to wait for job.
-            wait: Time in seconds between queries.
-
-        Returns:
-            OqtopusEstimationResult: the result of the estimation job.
-
-        Raises:
-            BackendError: If job cannot be found or if an authentication error occurred
-                or timeout occurs, etc.
-
-        """
-        if self._job.status not in JOB_FINAL_STATUS:
-            job = self.wait_for_completion(timeout, wait)
-            if job is None:
-                msg = f"Timeout occurred after {timeout} seconds."
-                raise BackendError(msg)
-            self._job = job
-        if self._job.status in {"failed", "cancelled"}:
-            msg = f"Job ended with status {self._job.status}."
-            raise BackendError(msg)
-
-        # edit json for OqtopusEstimationResult
-        result = self.job_info["result"]["estimation"]
-
-        return OqtopusEstimationResult(result)
-
-    def cancel(self) -> None:
-        """Cancel the job.
-
-        If the job statuses are success, failure, or cancelled,
-        then cannot be cancelled and an error occurs.
-
-        Raises:
-            BackendError: If job cannot be found or if an authentication error occurred
-                or if job cannot be cancelled, etc.
-
-        """
-        try:
-            self._job_api.cancel_job(self._job.job_id)
-            self.refresh()
-        except Exception as e:
-            msg = "To cancel job is failed."
-            raise BackendError(msg) from e
-
-    def to_json(self) -> str:
-        """Return a json string representation of the OqtopusEstimationJob.
-
-        Returns:
-            str: A json string representation of the OqtopusEstimationJob.
-
-        """
-        return json.dumps(self._job.to_dict(), cls=DateTimeEncoder)
-
-    def __repr__(self) -> str:
-        """Return a string representation of the OqtopusEstimationJob.
-
-        Returns:
-            str: A string representation of the OqtopusEstimationJob.
-
-        """
-        return self._job.to_str()
-
-
-class OqtopusEstimationBackend:
+class OqtopusEstimationBackend(OqtopusJobBackendBase):
     """A OQTOPUS backend for a estimation.
 
     Args:
@@ -407,35 +36,7 @@ class OqtopusEstimationBackend:
         self,
         config: OqtopusConfig | None = None,
     ) -> None:
-        super().__init__()
-
-        # set config
-        if config is None:
-            # if environment variables are set, use their values
-            url = os.getenv("OQTOPUS_URL")
-            api_token = os.getenv("OQTOPUS_API_TOKEN")
-            proxy = os.getenv("OQTOPUS_PROXY")
-            if url is not None and api_token is not None:
-                config = OqtopusConfig(
-                    url=url,
-                    api_token=api_token,
-                    proxy=proxy,
-                )
-            # load config from file
-            else:
-                config = OqtopusConfig.from_file()
-
-        # construct JobApi
-        rest_config = Configuration()
-        rest_config.host = config.url
-        if config.proxy:
-            rest_config.proxy = config.proxy
-        api_client = ApiClient(
-            configuration=rest_config,
-            header_name="q-api-token",
-            header_value=config.api_token,
-        )
-        self._job_api: JobApi = JobApi(api_client=api_client)
+        super().__init__(config=config)
 
     def estimate(  # noqa: PLR0917, PLR0913
         self,
@@ -583,28 +184,6 @@ class OqtopusEstimationBackend:
             response = self._job_api.get_job(response_submit_job.job_id)
         except Exception as e:
             msg = "To execute estimation on OQTOPUS Cloud is failed."
-            raise BackendError(msg) from e
-
-        return OqtopusEstimationJob(response, self._job_api)
-
-    def retrieve_job(self, job_id: str) -> OqtopusEstimationJob:
-        """Retrieve the job with the given id from OQTOPUS Cloud.
-
-        Args:
-            job_id: The id of the job to retrieve.
-
-        Returns:
-            The job with the given ``job_id``.
-
-        Raises:
-            BackendError: If job cannot be found or if an authentication error occurred,
-                etc.
-
-        """
-        try:
-            response = self._job_api.get_job(job_id)
-        except Exception as e:
-            msg = "To retrieve_job from OQTOPUS Cloud is failed."
             raise BackendError(msg) from e
 
         return OqtopusEstimationJob(response, self._job_api)
