+++ tests/quri_parts_oqtopus/backend/test_sampling.py
@@ -20,10 +20,12 @@ from quri_parts.backend import BackendError
 from quri_parts.circuit import QuantumCircuit
 
 from quri_parts_oqtopus.backend.config import OqtopusConfig
-from quri_parts_oqtopus.backend.sampling import (
-    OqtopusSamplingBackend,
+from quri_parts_oqtopus.backend.data.sampling import (
     OqtopusSamplingJob,
     OqtopusSamplingResult,
+)
+from quri_parts_oqtopus.backend.sampling import (
+    OqtopusSamplingBackend,
 )
 from quri_parts_oqtopus.rest import (
     JobApi,
